lookup for NS and A records
