+++ docker-compose.yaml
@@ -136,14 +136,6 @@ services:
     restart: always
     volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
 
-  karton-dns_resolver:
-    <<: *artemis-build-or-image
-    command: "python3 -m artemis.modules.dns_resolver"
-    depends_on: [karton-system, karton-logger]
-    env_file: .env
-    restart: always
-    volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
-  
   karton-dns_scanner:
     <<: *artemis-build-or-image
     command: "python3 -m artemis.modules.dns_scanner"
@@ -159,7 +151,15 @@ services:
     env_file: .env
     restart: always
     volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
-
+  
+  karton-domain_scanner:
+    <<: *artemis-build-or-image
+    command: "python3 -m artemis.modules.domain_scanner"
+    depends_on: [karton-system, karton-logger]
+    env_file: .env
+    restart: always
+    volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
+  
   karton-drupal_scanner:
     <<: *artemis-build-or-image
     command: "python3 -m artemis.modules.drupal_scanner"
@@ -183,7 +183,6 @@ services:
     env_file: .env
     restart: always
     volumes: ["./docker/karton.ini:/etc/karton/karton.ini", "${DOCKER_COMPOSE_ADDITIONAL_SHARED_DIRECTORY:-./shared}:/shared/"]
-
 
   karton-http_service_to_url:
     <<: *artemis-build-or-image
